langchain models/requirements.txt
huggingface_hub
langchain_huggingface
python-dotenv
langchain 
